created HistoricalWeatherHumidity:-model

+++ database/migrations/2023_02_02_094617_create_weather_forecast_models_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('weather_forecast_models', function (Blueprint $table) {
             $table->id();
-            $table->float('latitude',10,6);
-            $table->float('longitude', 10,6);
+            $table->decimal('latitude',14,10);
+            $table->decimal('longitude', 14,10);
             $table->dateTime('date_time_of_measurement');
-            $table->float('temperature');
+            $table->decimal('temperature');
             $table->timestamps();
 
             $table->unique(['latitude', 'longitude', 'date_time_of_measurement'], 'place_time_unique');
